python_client/requirements.txt
requests
python-dotenv
rich
typer[all]
